+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog 📋
 
+## [1.0.3] - 2025/08/29
+
+- Added authentication signature for API calls.
+- Added cancel order API calls.
+
 ## [1.0.2] - 2025/08/29
 
 - Working in the first strategy named MidCandle.